- Refactored to have example how subscriber would run - adding stub demo to see parts that would run on different servers for provider and subcriber
