+++ Oracle/Oracle.ts
