+++ Oracle/Responder.ts
