+++ Oracle/index.ts
@@ -1,0 +1,4 @@
+import {Oracle} from "./Oracle";
+
+const oracle = new Oracle()
+oracle.initialize().catch(console.error)
