+++ Oracle/types.ts
