+++ demo.ts
@@ -1,6 +1,12 @@
-import {Oracle} from "./src/Oracle";
+import {Oracle} from "./Oracle/Oracle";
 import {queryProvider} from "./subscriber/Susbcriber_example"
 
+/**
+ * This demo shows the simplest work flow
+ * - oracle can be run on oracle server,
+ * - oracle address then is given to subscriber
+ * - subscriber start query on their part and listen to responses from provider
+ */
 //to run on the oracle server
 const oracle = new Oracle()
 oracle.initialize().catch(console.error)
