+++ tsconfig.json
@@ -15,6 +15,6 @@
 		"strict": true
 	},
 	"include": [
-		"./src/**/*", "*.ts"
+		"./Oracle/**/*", "*.ts"
 	]
 }
